+++ examples/react-typescript/src/core/index.ts
@@ -1,3 +1,6 @@
-import { core } from './core';
+import { core } from "./core";
+import { App } from "./app";
+
+export { App };
 
 export default core;