+++ lib/state.ts
@@ -1,42 +1,52 @@
-import Dep from './dep';
-import Pulse from './pulse';
+import { Pulse, Dep } from './internal';
 import { copy, shallowmerge } from './utils';
 import { deepmerge } from './helpers/deepmerge';
-import { persistValue } from './storage';
 
 export class State<ValueType = any> {
+  // a unique key for this State
+  public name?: string;
+  // internal storage for the current value
   public _value: ValueType = null;
-  public set value(val: ValueType) {
-    this._value = val;
-  }
+  // getter for the current value
   public get value(): ValueType {
     if (this.instance().runtime.trackState) this.instance().runtime.foundState.add(this);
     return this._value;
   }
-  public dep: Dep = null;
-  public output?: any;
-  public watchers?: { [key: string]: any };
-  public previousState: ValueType = null;
-  public nextState: ValueType = null;
-  public isSet: boolean = false; // has been changed from initial value
-  public persistState: boolean;
-  public name?: string;
-  public typeOfVal?: string;
-  // sideEffects can be set by extended classes, such as Groups to build their output.
-  public sideEffects?: Function;
+  // dependency manager class
+  public dep: Dep;
 
+  // watchers
+  public watchers?: { [key: string]: any };
+  // the previous value of this State
+  public previousState: ValueType = null;
+  // a copy of the current value that will be used if no param is passed on State.set()
+  public nextState: ValueType = null;
+  // if the value has been changed from initial value
+  public isSet: boolean = false;
+  // should Pulse attempt to persist this State value
+  public persistState: boolean;
+  // if this State locked to a particular type
+  public typeOfVal?: string;
+  // for extended classes to perform actions upon state change
+  public sideEffects?: Function;
+  // for extended classes to store a derived value, such as Group
+  public output?: any;
+  // getter and setter for the State value, best for I/O binding
   public set bind(value: ValueType) {
     this.set(value);
   }
   public get bind(): ValueType {
     return this._value;
   }
+  // is value truthey or falsey
   public get exists(): boolean {
-    return !!this.value; // is value truthey or falsey
+    return !!this.value;
   }
 
   constructor(public instance: () => Pulse, public initialState, deps: Array<Dep> = []) {
+    // initialize the dependency manager
     this.dep = new Dep(deps);
+    // write the initial value to this State
     this.privateWrite(initialState);
   }
 
@@ -94,11 +104,14 @@ export class State<ValueType = any> {
 
   public persist(key?: string): this {
     this.persistState = true;
-    persistValue(this, key);
+    this.instance().storage.handleStatePersist(this, key);
     return this;
   }
 
-  // this creates a watcher that will fire a callback then destroy itself after invoking
+  /**
+   * @public
+   * Create a watcher that will fire a callback then destroy itself after invoking
+   */
   public onNext(callback: (value: ValueType) => void) {
     if (!this.watchers) this.watchers = {};
     this.watchers['_on_next_'] = () => {
@@ -107,19 +120,32 @@ export class State<ValueType = any> {
     };
   }
 
+  /**
+   * @public
+   * Set a name for this State, required to persist
+   */
   public key(key: string): this {
+    // if persist was attempted before, but no key was provided retry persist function
+    if (!this.name && this.persistState) this.persist(key);
     this.name = key;
     return this;
   }
 
-  public type(type: any): this {
+  /**
+   * @public
+   * Fix the type to one of 'String', 'Boolean', 'Array', 'Object' or 'Number'
+   */
+  public type(type: TypeString | TypeConstructor): this {
     const supportedConstructors = ['String', 'Boolean', 'Array', 'Object', 'Number'];
-    if (typeof type === 'function' && supportedConstructors.includes(type.name)) {
-      this.typeOfVal = type.name.toLowerCase();
-    }
+    if (typeof type === 'function' && supportedConstructors.includes(type.name)) this.typeOfVal = type.name.toLowerCase();
+    else if (typeof type === 'string' && supportedConstructors.map(i => i.toLowerCase()).includes(type)) this.typeOfVal = type.toLowerCase();
     return this;
   }
 
+  /**
+   * @public
+   * Watch state for changes, run callback on each change
+   */
   public watch(key: number | string, callback: (value: any) => void): this {
     if (!this.watchers) this.watchers = {};
 
@@ -130,15 +156,27 @@ export class State<ValueType = any> {
     return this;
   }
 
-  public undo() {
-    this.set(this.previousState);
-  }
-
+  /**
+   * @public
+   * Remove watcher by key
+   */
   public removeWatcher(key: number | string): this {
     delete this.watchers[key];
     return this;
   }
 
+  /**
+   * @public
+   * Restore previous state
+   */
+  public undo() {
+    this.set(this.previousState);
+  }
+
+  /**
+   * @public
+   * If State is boolean, invert
+   */
   public toggle(): this {
     if (typeof this._value === 'boolean') {
       // @ts-ignore
@@ -147,43 +185,75 @@ export class State<ValueType = any> {
     return this;
   }
 
+  /**
+   * @public
+   * Reset the State to as declared
+   */
   public reset(): this {
-    reset(this);
+    this.isSet = false;
+    this.previousState = null;
+    this.privateWrite(this.initialState);
+    if (this.persistState) this.instance().storage.remove(this.name);
     return this;
   }
 
-  // returns a fresh copy of the current value
+  /**
+   * @public
+   * Returns a copy of the current value, objects and arrays will be cloned
+   */
   public copy(): any {
     return copy(this.value);
   }
-
+  /**
+   * @public
+   * Is the value equal to parameter
+   */
   public is(x: any) {
     return this.value === x;
   }
 
+  /**
+   * @public
+   * Is the value not equal to parameter
+   */
   public isNot(x: any) {
     return this.value !== x;
   }
 
-  // INTERNAL
+  /**
+   * @internal
+   * Write value directly to State
+   */
   public privateWrite(value: any) {
     this._value = copy(value);
     this.nextState = copy(value);
-
+    // If
     if (this.persistState) this.instance().storage.set(this.name, this.getPersistableValue());
   }
 
+  /**
+   * @internal
+   *
+   */
   private isCorrectType(value): boolean {
     let type: string = typeof value;
     if (type === 'object' && Array.isArray(value)) type = 'array';
     return type === this.typeOfVal;
   }
 
+  /**
+   * @internal
+   *
+   */
   public destroy(): void {
     this.dep.deps.clear();
     this.dep.subs.clear();
   }
 
+  /**
+   * @internal
+   *
+   */
   public getPersistableValue(): any {
     return this.value;
   }
@@ -201,13 +271,9 @@ export const StateGroup = (instance: () => Pulse, stateGroup: Object): any => {
   }
   return group;
 };
+
 export default State;
 
-export function reset(instance: State) {
-  instance.isSet = false;
-  instance.previousState = null;
-  instance.privateWrite(instance.initialState);
-  if (instance.persistState) instance.instance().storage.remove(instance.name);
-}
-
 export type SetFunc<ValueType> = (state: ValueType) => ValueType;
+type TypeString = 'string' | 'boolean' | 'array' | 'object' | 'number';
+type TypeConstructor = StringConstructor | BooleanConstructor | ArrayConstructor | ObjectConstructor | NumberConstructor;