+++ lib/storage.ts
@@ -1,16 +1,16 @@
-import Pulse, { State } from './';
-import { defineConfig, isFunction, isAsync } from './utils';
+import { Pulse, State } from './internal';
+import { defineConfig, isAsync } from './utils';
 
 export interface StorageConfig {
-  type: 'custom' | 'localStorage';
-  prefix: string;
+  type?: 'custom' | 'localStorage';
+  prefix?: string;
   async?: boolean;
   get?: any;
   set?: any;
   remove?: any;
 }
 
-export default class Storage {
+export class Storage {
   public config: StorageConfig;
   private storageReady: boolean = false;
   public persistedState: Set<State> = new Set();
@@ -20,16 +20,13 @@ export default class Storage {
       prefix: 'pulse',
       type: 'localStorage'
     });
-
     // assume if user provided get, set or remove methods that the storage type is custom
-    if (this.config.get || this.config.set || this.config.remove) {
-      this.config.type = 'custom';
-    }
+    if (this.config.get || this.config.set || this.config.remove) this.config.type = 'custom';
 
-    if (this.localStorageAvailable() && this.config.type === 'localStorage') {
-      this.config.get = localStorage.getItem.bind(localStorage);
-      this.config.set = localStorage.setItem.bind(localStorage);
-      this.config.remove = localStorage.removeItem.bind(localStorage);
+    const ls = this.getLocalStorage();
+    if (this.config.type === 'localStorage' && ls) {
+      // assign localStorage crud functions to config object
+      ['get', 'set', 'remove'].forEach(type => (this.config[type] = ls[`${type}Item`].bind(ls)));
       this.storageReady = true;
     } else {
       // Local storage not available, fallback to custom.
@@ -40,7 +37,7 @@ export default class Storage {
         if (this.config.async === undefined && isAsync(this.config.get)) this.config.async = true;
         this.storageReady = true;
       } else {
-        console.warn('Pulse Error: Persistent storage not configured, check get, set and remove methods', this.config);
+        // console.warn('Pulse Error: Persistent storage not configured, check get, set and remove methods', this.config);
         this.storageReady = false;
       }
     }
@@ -55,7 +52,6 @@ export default class Storage {
           .then(res => {
             // if result is not JSON for some reason, return it.
             if (typeof res !== 'string') return resolve(res);
-
             resolve(JSON.parse(res));
           })
           .catch(reject);
@@ -82,43 +78,56 @@ export default class Storage {
   private getKey(key: string) {
     return `_${this.config.prefix}_${key}`;
   }
+  // used by State and Selector to persist value inside storage
 
-  private localStorageAvailable() {
+  public handleStatePersist(state: State, key: string) {
+    const storage = this;
+    // validation
+    if (!key && state.name) {
+      key = state.name;
+    } else if (!key) {
+      return;
+      // console.warn('Pulse Persist Error: No key provided');
+    } else {
+      state.name = key;
+    }
+    // add ref to state instance inside storage
+    storage.persistedState.add(state);
+
+    // handle the value
+    const handle = (storageVal: any) => {
+      // if no storage value found, set current value in storage
+      if (storageVal === null) storage.set(state.name, state.getPersistableValue());
+      // if Selector, select current storage value
+      else if (typeof state['select'] === 'function' && (typeof storageVal === 'string' || typeof storageVal === 'number'))
+        state['select'](storageVal);
+      // otherwise just ingest the storage value so that the State updates
+      else state.instance().runtime.ingest(state, storageVal);
+    };
+    // Check if promise, then handle value
+    if (storage.config.async) storage.get(state.name).then((value: any) => handle(value));
+    // non promise
+    else handle(storage.get(state.name));
+  }
+
+  private getLocalStorage() {
     try {
-      localStorage.setItem('_', '_');
-      localStorage.removeItem('_');
-      return true;
+      const ls = window?.localStorage ? window.localStorage : localStorage;
+      if (typeof ls.getItem !== 'function') return false;
+      return ls;
     } catch (e) {
       return false;
     }
   }
 }
 
-// used by State and Selector to persist value inside storage
-export function persistValue(state: State, key: string) {
-  const storage = state.instance().storage;
-  // validation
-  if (!key && state.name) {
-    key = state.name;
-  } else if (!key) {
-    console.warn('Pulse Persist Error: No key provided');
-  } else {
-    state.name = key;
-  }
-  // add ref to state instance inside storage
-  storage.persistedState.add(state);
+function isFunction(func: () => any) {
+  return typeof func === 'function';
+}
 
-  // handle the value
-  const handle = (storageVal: any) => {
-    // if no storage value found, set current value in storage
-    if (storageVal === null) storage.set(state.name, state.getPersistableValue());
-    // if Selector, select current storage value
-    else if (typeof state['select'] === 'function' && (typeof storageVal === 'string' || typeof storageVal === 'number')) state['select'](storageVal);
-    // otherwise just ingest the storage value so that the State updates
-    else state.instance().runtime.ingest(state, storageVal);
-  };
-  // Check if promise, then handle value
-  if (storage.config.async) storage.get(state.name).then((value: any) => handle(value));
-  // non promise
-  else handle(storage.get(state.name));
+export default Storage;
+
+// Handy utils
+export function persist(items: Array<State>): void {
+  items.forEach(item => item.persist(item.name));
 }